# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Kinoko API
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /texts:
    get:
      operationId: listTexts
      tags: [texts]
      security: []
      parameters:
        - name: If-None-Match
          in: header
          schema:
            type: string
      responses:
        "200":
          description: OK
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextsCatalogue"
        "304":
          description: Not Modified
        "422":
          $ref: "#/components/responses/Validation"
  /texts/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: putText
      tags: [texts]
      x-permission: text.edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextWrite"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextEntry"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"

    delete:
      operationId: resetText
      tags: [texts]
      x-permission: text.edit
      parameters:
        - name: locale
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [access]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/species-counts:
    get:
      operationId: getAdminSpeciesCounts
      tags: [access]
      x-permission: species.edit
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SpeciesCountsEntry"
                required: [items]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /permissions:
    get:
      operationId: listPermissions
      tags: [access]
      x-permission: role.manage
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/PermissionEntry"
                required: [items]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /roles:
    get:
      operationId: listRoles
      tags: [access]
      x-permission: role.manage
      parameters:
        - $ref: "#/components/parameters/Limit50"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"

    post:
      operationId: createRole
      tags: [access]
      x-permission: role.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "422":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /roles/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getRole
      tags: [access]
      x-permission: role.manage
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"

    patch:
      operationId: updateRole
      tags: [access]
      x-permission: role.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"

    delete:
      operationId: deleteRole
      tags: [access]
      x-permission: role.manage
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"
  /people:
    get:
      operationId: listPeople
      tags: [access]
      x-permission: role.assign
      parameters:
        - $ref: "#/components/parameters/Limit50"
        - $ref: "#/components/parameters/Cursor"
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"
  /people/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getPerson
      tags: [access]
      x-permission: role.assign
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"

    delete:
      operationId: deletePerson
      tags: [access]
      x-permission: role.assign
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"
  /people/{id}/roles:
    parameters:
      - $ref: "#/components/parameters/Id"
    put:
      operationId: setPersonRoles
      tags: [access]
      x-permission: role.assign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
                    format: uuid
              required: [roleIds]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Validation"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit50:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 50
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
  responses:
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Validation:
      description: Unprocessable Entity
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          description: Machine-readable key, for example not_found or last_admin.
        detail:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
      required: [type, title, status, code]
    FieldError:
      type: object
      properties:
        field:
          type: string
        code:
          type: string
      required: [field, code]
...
    Permission:
      type: string
      enum:
        - species.edit
        - image.review
        - image.submit
        - find.review
        - role.manage
        - role.assign
        - text.edit
        - run.manage
        - group.manage
    Area:
      type: string
      enum: [species, interface, access, data]
    SpeciesCountsEntry:
      type: object
      properties:
        speciesId:
          type: string
          format: uuid
        records:
          type: integer
        finds:
          type: integer
        photos:
          type: integer
      required: [speciesId, records, finds, photos]
    AdminSummary:
      description: Zähler der Verwaltungsübersicht. Ein Feld fehlt ohne das Recht dazu.
      type: object
      properties:
        texts:
          type: integer
        photos:
          type: integer
        photosPending:
          type: integer
        species:
          type: integer
        roles:
          type: integer
        permissions:
          type: integer
        people:
          type: integer
        finds:
          type: integer
        findsPending:
          type: integer
        runs:
          type: integer
        runsRunning:
          type: integer
    PermissionEntry:
      type: object
      properties:
        key:
          $ref: "#/components/schemas/Permission"
        area:
          $ref: "#/components/schemas/Area"
      required: [key, area]
...
    TextWrite:
      type: object
      properties:
        locale:
          type: string
        value:
          type: string
          minLength: 1
          maxLength: 2000
      required: [locale, value]
    TextEntry:
      type: object
      properties:
        key:
          type: string
        values:
          type: object
          additionalProperties:
            type: string
        changed:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required: [key, values, changed, updatedAt]
    TextsCatalogue:
      type: object
      properties:
        revision:
          type: string
        locales:
          type: array
          items:
            type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TextEntry"
      required: [revision, locales, entries]
    RoleBrief:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
      required: [id, slug, name]
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          oneOf:
            - type: string
            - type: "null"
        builtIn:
          type: boolean
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        peopleCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - builtIn
        - permissions
        - peopleCount
        - createdAt
        - updatedAt
    RoleCreate:
      type: object
      properties:
        slug:
          type: string
          pattern: "^[a-z][a-z0-9-]*$"
        name:
          type: string
        description:
          type: string
          maxLength: 200
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
      required: [slug, name]
    RoleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          oneOf:
            - type: string
            - type: "null"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
    RolePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        nextCursor:
          oneOf:
            - type: string
            - type: "null"
      required: [items, nextCursor]
...
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sub:
          type: string
        email:
          oneOf:
            - type: string
            - type: "null"
        name:
          oneOf:
            - type: string
            - type: "null"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleBrief"
        createdAt:
          type: string
          format: date-time
      required: [id, sub, roles, createdAt]
    PersonPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        nextCursor:
          oneOf:
            - type: string
            - type: "null"
      required: [items, nextCursor]
...